+++ GestiNotes3.0/src/vue/ElevePanel.java
@@ -151,7 +151,7 @@ public class ElevePanel extends UtilePanel {
             eleve.setDateNaissance(getChamp(2).getText());
             ArrayList<Evaluation> evaluations = eleve.getTabEvaluation();
             for (int i = 0; i < NBR_NOTES; i++) {
-                evaluations.get(i).setNote(Double.parseDouble(getChamp(i + 4).getText()));
+                evaluations.get(i).setNote(Float.parseFloat(getChamp(i + 4).getText()));
             }
             notification("Modification effectue");
